Nagivation Bar changed

+++ django_app/users/views.py
@@ -37,7 +37,7 @@ def profile(request):
 		if u_form.is_valid() and p_form.is_valid():
 			u_form.save()
 			p_form.save()
-			os.remove(prev_imagepath)
+			if("default" not in prev_imagepath): os.remove(prev_imagepath)
 			messages.success(request, f'Your account has been updated!!')
 			return redirect('profile')
 		else:
